statPage data 훅 분리
