+++ src/stats/components/StatsPage.tsx
@@ -1,100 +1,31 @@
-import { useQueryClient } from "@tanstack/react-query"
 import { useState } from "react"
-import { useRegisterTabHandler } from "@/shared/contexts/BottomTabHandlerContext"
 import { usePerformanceMonitoring } from "@/shared/hooks/usePerformanceMonitoring"
-import { useRemoteConfig } from "@/shared/hooks/useRemoteConfig"
 import { ScrollArea } from "@/shared/ui/scroll-area"
 import StatsHeader from "@/stats/components/StatsHeader"
 import { StatsNoticeBanner } from "@/stats/components/StatsNoticeBanner"
-import { useWritingStatsV2 } from "@/stats/hooks/useWritingStatsV2"
-import { useUserInBoard } from "@/user/hooks/useUserInBoard"
-import { useNavigate } from "react-router-dom"
-import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/shared/ui/tabs"
-import { useCommentingStats } from "@/stats/hooks/useCommentingStats"
+import { useStatsPageData } from "@/stats/hooks/useStatsPageData"
 import { UserPostingStatsCardList } from "@/stats/components/UserPostingStatsCardList"
 import { UserCommentStatsCardList } from "@/stats/components/UserCommentStatsCardList"
 import React from "react"
 import { Loader2 } from "lucide-react"
+import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/shared/ui/tabs"
+import { useNavigate } from "react-router-dom"
 
 // 통계 페이지 스크롤 영역의 고유 ID
 const STATS_SCROLL_ID = 'stats-scroll';
 
 type TabType = 'posting' | 'commenting';
 
-function useStatsPageData(tab: TabType) {
-    const queryClient = useQueryClient();
-    const navigate = useNavigate();
-    
-    // Remote Config에서 활성 게시판 ID 가져오기
-    const { 
-        value: activeBoardId, 
-        isLoading: isLoadingConfig,
-        error: configError
-    } = useRemoteConfig<string>(
-        'active_board_id', 
-        '5rfpfRBuhRFZB13dJVy8' // 기본값을 문자열로 변경
-    );
-    
-    // 활성 게시판 권한이 있는 사용자 가져오기
-    const { users: activeUsers, isLoading: isLoadingUsers, error: usersError } = useUserInBoard(
-      activeBoardId ? [activeBoardId] : []
-    );
-
-    // 사용자 ID 배열로 통계 가져오기
-    const { 
-        data: writingStats, 
-        isLoading: isLoadingStats, 
-        error: statsError 
-    } = useWritingStatsV2(
-        activeUsers.map((user) => user.uid)
-    );
-
-    const { 
-        data: commentingStats, 
-        isLoading: isLoadingCommenting, 
-        error: commentingError 
-    } = useCommentingStats(
-        activeUsers.map((user) => user.uid)
-    );
-    
-    const isLoading = isLoadingConfig || isLoadingUsers || (tab === 'posting' ? isLoadingStats : isLoadingCommenting);
-    const error = configError || usersError || (tab === 'posting' ? statsError : commentingError);
-    
-    // 통계 새로고침 핸들러
-    const handleRefreshStats = () => {
-        window.scrollTo({ top: 0, behavior: 'smooth' });
-        
-        // 2. 통계 관련 쿼리 캐시 무효화
-        queryClient.invalidateQueries(['activeUsers', activeBoardId]);
-        queryClient.invalidateQueries(['writingStatsV2']);
-        queryClient.invalidateQueries(['commentingStats']);
-    };
-    
-    // Stats 탭 핸들러 등록
-    useRegisterTabHandler('Stats', handleRefreshStats);
-    
-    return {
-        activeUsers,
-        writingStats,
-        commentingStats,
-        isLoading,
-        error,
-        handleRefreshStats,
-        navigate,
-        isLoadingCommenting
-    };
-}
-
 export default function StatsPage() {
     usePerformanceMonitoring('StatsPage');
+    const navigate = useNavigate();
     const [tab, setTab] = useState<TabType>('posting');
+    
     const {
         writingStats,
         commentingStats,
         isLoading,
         error,
-        handleRefreshStats,
-        navigate,
         isLoadingCommenting
     } = useStatsPageData(tab);
